+++ reference-implementation/unittests/tosa.cpp
@@ -1018,7 +1018,7 @@ TEST(tosa, pad) {
 TEST(tosa, tile) {
   { // 1d case
     Tensor1D<int32_t, 4> input{1, 2, 3, 4};
-    Tensor1D<int64_t, 1> multiples{3};
+    Tensor1D<int32_t, 1> multiples{3};
     Tensor1D<int32_t, 12> expected_result{1, 2, 3, 4, 1, 2, 3, 4, 1, 2, 3, 4};
     Tensor1D<int32_t, 12> result =
         tosa::tile<Tensor1D<int32_t, 12>>(input, multiples);
@@ -1026,7 +1026,7 @@ TEST(tosa, tile) {
   }
   { // 2d case
     Tensor2D<int32_t, 3, 3> input{1, 2, 3, 4, 5, 6, 7, 8, 9};
-    Tensor1D<int64_t, 2> multiples{2, 3};
+    Tensor1D<int32_t, 2> multiples{2, 3};
     Tensor2D<int32_t, 6, 9> expected_result{
         1, 2, 3, 1, 2, 3, 1, 2, 3, 4, 5, 6, 4, 5, 6, 4, 5, 6,
         7, 8, 9, 7, 8, 9, 7, 8, 9, 1, 2, 3, 1, 2, 3, 1, 2, 3,
@@ -1037,7 +1037,7 @@ TEST(tosa, tile) {
   }
   { // 2d case
     Tensor2D<int32_t, 2, 4> input{1, 2, 3, 4, 5, 6, 7, 8};
-    Tensor1D<int64_t, 2> multiples{3, 1};
+    Tensor1D<int32_t, 2> multiples{3, 1};
     Tensor2D<int32_t, 6, 4> expected_result{1, 2, 3, 4, 5, 6, 7, 8, 1, 2, 3, 4,
                                             5, 6, 7, 8, 1, 2, 3, 4, 5, 6, 7, 8};
     Tensor2D<int32_t, 6, 4> result =
@@ -1046,7 +1046,7 @@ TEST(tosa, tile) {
   }
   { // 3d case
     Tensor3D<int32_t, 2, 2, 3> input{1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12};
-    Tensor1D<int64_t, 3> multiples{2, 1, 3};
+    Tensor1D<int32_t, 3> multiples{2, 1, 3};
     Tensor3D<int32_t, 4, 2, 9> expected_result{
         1, 2, 3, 1, 2, 3, 1, 2, 3, 4,  5,  6,  4,  5,  6,  4,  5,  6,
         7, 8, 9, 7, 8, 9, 7, 8, 9, 10, 11, 12, 10, 11, 12, 10, 11, 12,
@@ -1058,7 +1058,7 @@ TEST(tosa, tile) {
   }
   {
     Tensor3D<int32_t, 1, 3, 2> input{1, 2, 3, 4, 5, 6};
-    Tensor1D<int64_t, 3> multiples{2, 2, 2};
+    Tensor1D<int32_t, 3> multiples{2, 2, 2};
     Tensor3D<int32_t, 2, 6, 4> expected_result{
         1, 2, 1, 2, 3, 4, 3, 4, 5, 6, 5, 6, 1, 2, 1, 2, 3, 4, 3, 4, 5, 6, 5, 6,
         1, 2, 1, 2, 3, 4, 3, 4, 5, 6, 5, 6, 1, 2, 1, 2, 3, 4, 3, 4, 5, 6, 5, 6};
@@ -1069,7 +1069,7 @@ TEST(tosa, tile) {
   { // 4d case
     Tensor4D<int32_t, 2, 2, 2, 2> input{1, 2,  3,  4,  5,  6,  7,  8,
                                         9, 10, 11, 12, 13, 14, 15, 16};
-    Tensor1D<int64_t, 4> multiples{2, 3, 1, 2};
+    Tensor1D<int32_t, 4> multiples{2, 3, 1, 2};
     Tensor4D<int32_t, 4, 6, 2, 4> expected_result{
         1,  2,  1,  2,  3,  4,  3,  4,  5,  6,  5,  6,  7,  8,  7,  8,  1,  2,
         1,  2,  3,  4,  3,  4,  5,  6,  5,  6,  7,  8,  7,  8,  1,  2,  1,  2,
